adding getalltrades && delete trade

+++ src/main/java/marcowidesott/CapstoneBE/entities/Trade.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.UUID;
@@ -39,8 +40,8 @@ public class Trade {
 
     @Enumerated(EnumType.STRING)
     private TradeResult result;  // Esito del trade (Profitto, Perdita, Break Even)
-    
-    private Double profitLoss;
+
+    private BigDecimal profitLoss;
 
     @Enumerated(EnumType.STRING)
     private Asset asset;
